About me :)
I am a proactive person, both a good communicator and a good listener. I am a sales person, so I am focused on performing results, I like to learn and constantly develop myself and I hope that together with you who are reading this description, we can form a team that we can be proud of.